+++ includes/lexing.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   lexing.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: fcaldas- <fcaldas-@student.42.fr>          +#+  +:+       +#+        */
+/*   By: edcastro <edcastro@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/22 13:17:57 by educastro         #+#    #+#             */
-/*   Updated: 2024/07/23 13:10:30 by fcaldas-         ###   ########.fr       */
+/*   Updated: 2024/07/24 16:36:55 by edcastro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,29 +58,35 @@ typedef struct			s_aux_token
 }	t_aux_token;
 
 // get_states.c
-int				token_state_is_final(int state);
 int				get_token_type(int state);
-int				token_state_requires_backtrack(int state);
 int				token_get_next_state(int state, char c);
+int				token_state_requires_backtrack(int state);
+int				token_state_is_final(int state);
+
+// token_states_1.c
 int				get_state_1(char c);
 int				get_state_40(char c);
 int				get_state_50(char c);
 int				get_state_60(char c);
 int				get_state_70(char c);
+
+// token_states_2.c
 int				get_state_80(char c);
 int				get_state_81(char c);
 int				get_state_82(char c);
-int				get_next_state(int state, char c);
-
 
 // tokenizer.c
 t_token_list	*get_token_list(char *str);
 t_token_list	*token_get_sublist(t_token_list *token_lst, int start, int lst_len);
 void			token_final_state(t_aux_token *aux, t_token_list **token_list, char *str);
 
-
 // utils.c
+void	token_clear_list(t_token_list **token_list);
 int		is_whitespace(char c);
 int		is_metacharacter(char c);
+
+// tests.c (TEMPORARY)
+const char *get_token_type_string(enum e_token type);
+void print_token_list(t_token_list *token_list);
 
 #endif
